Fix : Update Unused Picture Delete - User Specific
Article에 사용되지 않은 Picture를 삭제할 때 요청을 보낸 사용자가 작성한것만 삭제 되도록 수정했습니다.

+++ ai_process/views.py
@@ -58,7 +58,7 @@ class PicgenView(APIView):
 
         post요청 시 입력받은 사진으로 변환된 사진을 생성하여 반환합니다.
         """
-        Picture.objects.filter(article=None).delete()
+        Picture.objects.filter(article=None, author=request.user).delete()
         serializer = PictureSerializer(data=request.data)
         if serializer.is_valid():
             orm = serializer.save()
